requirements.txt
blinker==1.8.2
branca==0.7.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
Flask-Cors==5.0.0
folium==0.17.0
fonttools==4.54.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
rdp==0.8
requests==2.32.3
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
xyzservices==2024.9.0
requests_html
lxml_html_clean
python-dotenv